chore: add segmented control tooltip
